+++ generator/generator.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from data.task import Task
+from data.Task import Task
 
 # Generates valid tasks list (no collisions)
 # Collisions will be added later
